@@ -1,8 +1,8 @@
 {
   "name": "react-native-super-chooser",
-  "version": "0.0.3",
+  "version": "0.0.5",
   "description": "a chooser of react-native",
-  "main": "index.js",
+  "main": "./lib/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -18,10 +18,10 @@
   ],
   "author": "senze.fan",
   "license": "MIT",
-  "dependencies": {
+  "peerDependencies": {
     "prop-types": "^15.5.10",
-    "react": "^15.6.1",
-    "react-native": "^0.48.2"
+    "react": "16.0.0-alpha.12",
+    "react-native": ">=0.47.0"
   },
   "bugs": {
     "url": "https://github.com/fansenze/react-native-super-chooser/issues"
