@@ -1,1 +1,9 @@
-module.exports = require('./lib')
+import SuperChooser from './lib'
+import Symbol from './lib/symbol'
+
+export {
+  SuperChooser,
+  Symbol
+}
+
+export default SuperChooser
